requirements_rag.txt
# Core RAG Application Dependencies

# Web Framework
fastapi==0.116.1
uvicorn==0.35.0

# Vector Database
pinecone-client==3.2.2

# Embeddings & AI
sentence-transformers==2.2.2
cerebras-cloud-sdk==1.50.1

# Document Processing
langchain==0.0.350
langchain-community==0.0.13

# Utilities
huggingface-hub==0.20.3
python-dotenv==1.0.0